requirements.txt
pandas
numpy
PyYAML
typer
matplotlib
pytest
pytz
python-dateutil